chore: Improve test infra: MySQL options, cancellation, timeouts
- Check cancellation token before DB connect in EntityFrameworkInitializer for better test responsiveness.
- Add connection options to MySqlContainerFixture for improved MySQL compatibility.
- Move MySqlDatabaseServiceFixture to its own file (no logic changes).
- Increase OutboxTestsBase timeout to 5 minutes for slow container startups.
