+++ tests/NetEvolve.Pulse.Tests.Integration/Outbox/OutboxTestsBase.cs
@@ -9,7 +9,7 @@ using NetEvolve.Pulse.Outbox;
 using NetEvolve.Pulse.Tests.Integration.Internals;
 
 [TestGroup("Outbox")]
-[Timeout(60_000)] // Increased timeout to accommodate potential delays in CI environments, especially when using SQL Server containers.
+[Timeout(300_000)] // Increased timeout to accommodate potential delays in CI environments, especially when using SQL Server or MySQL containers that can take a long time to cold-start.
 public abstract class OutboxTestsBase(
     IDatabaseServiceFixture databaseServiceFixture,
     IDatabaseInitializer databaseInitializer
